Add User handling for to dos

+++ backend/src/main/java/com/example/todo/service/ToDoService.java
@@ -1,5 +1,6 @@
 package com.example.todo.service;
 
+import com.example.todo.model.User;
 import com.example.todo.model.ToDo;
 import com.example.todo.model.ValidationResult;
 import com.example.todo.repository.ToDoRepository;
@@ -10,19 +11,24 @@ import org.springframework.stereotype.Service;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
 public class ToDoService {
     private final ToDoRepository toDoRepository;
+    private final UserService userService;
 
     public List<ToDo> getToDos() {
-        List<ToDo> toDos = this.toDoRepository.findAll();
+        User currentUser = this.userService.getCurrentUser();
+        List<ToDo> toDos = this.toDoRepository.findAllByUser(currentUser);
         this.sortToDosByDeadline(toDos);
         return toDos;
     }
 
     public ValidationResult saveToDo(final ToDo toDo) {
+        User currentUser = this.userService.getCurrentUser();
+        toDo.setUser(currentUser);
         ValidationResult validationResult = this.validateToDo(toDo);
         if (validationResult.isNotValid()) {
             return validationResult;
@@ -33,11 +39,20 @@ public class ToDoService {
 
     public void toggleDone(final Long id) {
         ToDo toDo = this.toDoRepository.getReferenceById(id);
+        User currentUser = this.userService.getCurrentUser();
+        if (!Objects.equals(currentUser.getId(), toDo.getUser().getId())) {
+            return;
+        }
         toDo.setDone(!toDo.isDone());
         this.toDoRepository.save(toDo);
     }
 
     public void deleteToDo(final Long id) {
+        User currentUser = this.userService.getCurrentUser();
+        ToDo toDo = this.toDoRepository.findById(id).orElseThrow();
+        if (!Objects.equals(currentUser.getId(), toDo.getUser().getId())) {
+            return;
+        }
         this.toDoRepository.deleteById(id);
     }
 
@@ -62,7 +77,7 @@ public class ToDoService {
         return (toDo.getDeadline() != null && toDo.getDeadline().before(today));
     }
 
-    public void sortToDosByDeadline(List<ToDo> toDos) {
+    private void sortToDosByDeadline(List<ToDo> toDos) {
         toDos.sort((toDo1, toDo2) -> {
             if (toDo1.getDeadline() == null && toDo2.getDeadline() == null) {
                 return 0;
@@ -76,4 +91,5 @@ public class ToDoService {
             return toDo1.getDeadline().compareTo(toDo2.getDeadline());
         });
     }
+
 }
